requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# AI / LLM
anthropic>=0.25.0

# RAG / Vector store
chromadb>=0.5.0

# Configuration
python-dotenv>=1.0.0

# Validation (FastAPI already pulls this in, pinned for clarity)
pydantic>=2.7.0

# HTTP client (useful for testing endpoints)
httpx>=0.27.0

# Required by FastAPI for form/multipart parsing
python-multipart>=0.0.9
